[Keyboard] Add Z60 Re-organized keyboard directory
Nothing is changed except that the files been moved into 'solder' & 'hotswap' subfolders.
