+++ keyboards/quarkeys/z60/readme.md
@@ -1,0 +1,27 @@
+# Quarkeys Z60 Keyboard
+
+![Z60](https://static.wixstatic.com/media/72b918_45f82258186646d4bfa57c9d73ccc9bc~mv2.jpeg)
+![Layout](https://static.wixstatic.com/media/72b918_bb37ec16543545b0aac331d5a7fa977c~mv2.png)
+
+A 60% Gasket keyboard with innovated top enclosure made by Quarkeys, second of out projects.
+
+* Keyboard Maintainer: [Quarkeys Studio](www.quarkeys.com)
+* Hardware Supported: Z60-ACR, Z60-Evolve
+* Hardware Specs: Solder version with per-key RGBs, layer & Capslock indicators
+                  Hotswap version with per-key RGBs, layer & Capslock indicators
+
+Make example for this keyboard solder version keymap(after setting up your build environment):
+
+    make quarkeys/z60/solder:default
+
+Make example for this keyboard hotswap version keymap:
+
+    make quarkeys/z60/hotswap:default
+
+Flashing example for this keyboard:
+
+    make quarkeys/z60/solder:default:flash
+
+**Bootloader:** Press the physical reset button (ECS position of layer 1)
+
+See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
