+++ keyboards/quarkeys/z60/solder/info.json
