+++ keyboards/quarkeys/z60/solder/solder.c
@@ -12,4 +12,4 @@
 / along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "z60.h"
+#include "solder.h"
