--- keyboards/quarkeys/z67sd/info.json
@@ -1,92 +1,0 @@
-{
-    "keyboard_name": "Z67 Solder", 
-    "manufacturer": "Quarkeys Stuidio",
-    "url": "www.quarkeys.com", 
-    "maintainer": "TommyZ", 
-    "usb": {
-        "vid": "0x8490",
-        "pid": "0x4101",
-        "device_version": "0.0.1"
-    },
-    "layouts": {
-        "LAYOUT": {
-            "layout": [
-                {"label":"Esc", "x":0, "y":0}, 
-                {"label":"!", "x":1, "y":0}, 
-                {"label":"@", "x":2, "y":0}, 
-                {"label":"#", "x":3, "y":0}, 
-                {"label":"$", "x":4, "y":0}, 
-                {"label":"%", "x":5, "y":0}, 
-                {"label":"^", "x":6, "y":0}, 
-                {"label":"&", "x":7, "y":0}, 
-                {"label":"*", "x":8, "y":0}, 
-                {"label":"(", "x":9, "y":0}, 
-                {"label":")", "x":10, "y":0}, 
-                {"label":"_", "x":11, "y":0}, 
-                {"label":"+", "x":12, "y":0}, 
-                {"label":"Back", "x":13, "y":0}, 
-                {"label":"~", "x":14, "y":0}, 
-                {"label":"Delete", "x":15, "y":0}, 
-                
-                {"label":"Tab", "x":0, "y":1, "w":1.5}, 
-                {"label":"Q", "x":1.5, "y":1},
-                 {"label":"W", "x":2.5, "y":1}, 
-                 {"label":"E", "x":3.5, "y":1}, 
-                 {"label":"R", "x":4.5, "y":1},
-                 {"label":"T", "x":5.5, "y":1}, 
-                 {"label":"Y", "x":6.5, "y":1}, 
-                 {"label":"U", "x":7.5, "y":1}, 
-                 {"label":"I", "x":8.5, "y":1}, 
-                 {"label":"O", "x":9.5, "y":1}, 
-                 {"label":"P", "x":10.5, "y":1}, 
-                 {"label":"{", "x":11.5, "y":1}, 
-                 {"label":"}", "x":12.5, "y":1}, 
-                 {"label":"|", "x":13.5, "y":1, "w":1.5}, 
-                 {"label":"PgUp", "x":15, "y":1}, 
-                 
-                 {"label":"Caps Lock", "x":0, "y":2, "w":1.75}, 
-                 {"label":"A", "x":1.75, "y":2}, 
-                 {"label":"S", "x":2.75, "y":2}, 
-                 {"label":"D", "x":3.75, "y":2}, 
-                 {"label":"F", "x":4.75, "y":2}, 
-                 {"label":"G", "x":5.75, "y":2}, 
-                 {"label":"H", "x":6.75, "y":2}, 
-                 {"label":"J", "x":7.75, "y":2}, 
-                 {"label":"K", "x":8.75, "y":2}, 
-                 {"label":"L", "x":9.75, "y":2}, 
-                 {"label":":", "x":10.75, "y":2}, 
-                 {"label":"\"", "x":11.75, "y":2}, 
-                 {"label":"~", "x":12.75, "y":2}, 
-                 {"label":"Enter", "x":13.75, "y":2, "w":1.25}, 
-                 {"label":"PgDn", "x":15, "y":2}, 
-                 
-                 {"label":"Shift", "x":0, "y":3, "w":1.25}, 
-                 {"label":"|", "x":1.25, "y":3}, 
-                 {"label":"Z", "x":2.25, "y":3}, 
-                 {"label":"X", "x":3.25, "y":3}, 
-                 {"label":"C", "x":4.25, "y":3}, 
-                 {"label":"V", "x":5.25, "y":3}, 
-                 {"label":"B", "x":6.25, "y":3}, 
-                 {"label":"N", "x":7.25, "y":3}, 
-                 {"label":"M", "x":8.25, "y":3}, 
-                 {"label":"<", "x":9.25, "y":3}, 
-                 {"label":">", "x":10.25, "y":3}, 
-                 {"label":"?", "x":11.25, "y":3}, 
-                 {"label":"Shift", "x":12.25, "y":3, "w":1.75}, 
-                 {"label":"\u2191", "x":14, "y":3}, 
-                 {"label":"Fn", "x":15, "y":3}, 
-                 
-                 {"label":"Ctrl", "x":0, "y":4, "w":1.25}, 
-                 {"label":"Win", "x":1.25, "y":4, "w":1.25}, 
-                 {"label":"Alt", "x":2.5, "y":4, "w":1.25}, 
-                 {"x":3.75, "y":4, "w":6.25}, 
-                 {"label":"Alt", "x":10, "y":4}, 
-                 {"label":"Win", "x":11, "y":4}, 
-                 {"label":"Ctrl", "x":12, "y":4}, 
-                 {"label":"\u2190", "x":13, "y":4}, 
-                 {"label":"\u2193", "x":14, "y":4}, 
-                 {"label":"\u2192", "x":15, "y":4}
-            ]
-        }
-    }
-}
