--- keyboards/quarkeys/z67sd/readme.md
@@ -1,26 +1,0 @@
-# Quarkeys Z67 Solder PCB
-
-![Z67](https://static.wixstatic.com/media/72b918_c3a6848a520f47888bbcb8f4a8019999~mv2.png)
-![Layout](https://static.wixstatic.com/media/72b918_8e66a20d353547fca89a3413336cd688~mv2.png)
-
-A 65% Gasket keyboard made by Quarkeys, third of out project.
-
-* Keyboard Maintainer: [Quarkeys Studio](www.quarkeys.com)
-* Hardware Supported: Z67 Solder PCB
-* Hardware Specs: 
-
-Make example for this keyboard default keymap(after setting up your build environment):
-
-    make quarkeys/z67sd:default
-
-Make example for this keyboard keymap for via:
-
-    make quarkeys/z67sd:via
-
-Flashing example for this keyboard:
-
-    make quarkeys/z67sd:default:flash
-
-**Bootloader:** Press the physical reset button (ECS position of layer 1)
-
-See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
